Remove link to 'The Erlang Rationale'
The link is broken. Correspondence with the author (Robert Virding,
2014-06-17) suggested that the name is now mainly used as a conference
presentation about historical aspects of Erlang.

The interesting part in the original link was the section on
the IO system and on process groups, Jobs and JCL. This isn't
really covered anywhere else as far as I know.

As at this commit, the original document is still to be found
here: http://vidiowiki.com/media/paper/0m-V9dOL0AeY%20The_Erlang_Rationale.pdf