Remove defunct contributors section
The whole Mnesia section was written by Chris Pressey, Ulf Wiger
and Sean Hinde, but that was a long time ago and now part has been
updated by Magnus Henoch. Instead of tracking contributions in
the FAQ text, I've removed it and if people are curious 'git' has
a record.